+++ src/main/java/roomescape/dao/JdbcReservationTimeDao.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
-import roomescape.domain_entity.ReservationTime;
+import roomescape.domain.ReservationTime;
 import roomescape.mapper.ReservationTimeMapper;
 
 @Component
