+++ src/main/java/roomescape/dao/ThemeDao.java
@@ -1,7 +1,7 @@
 package roomescape.dao;
 
 import java.util.List;
-import roomescape.domain_entity.Theme;
+import roomescape.domain.Theme;
 
 public interface ThemeDao {
