+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.domain_entity;
+package roomescape.domain;
 
 import java.time.LocalDate;
