+++ src/main/java/roomescape/dto/ThemeRequest.java
@@ -1,6 +1,6 @@
 package roomescape.dto;
 
-import roomescape.domain_entity.Theme;
+import roomescape.domain.Theme;
 
 public record ThemeRequest(
         String name, String description, String thumbnail
