+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -1,8 +1,7 @@
-package roomescape.domain_entity;
+package roomescape.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
 import java.time.LocalTime;
 import java.time.temporal.ChronoUnit;
 import java.util.stream.Stream;
